@@ -7,31 +7,61 @@ using System.Runtime.InteropServices.Marshalling;
 using System.Runtime.Intrinsics;
 using System.Text;
 
+using static SharpGit2.GitNativeApi;
+
 namespace SharpGit2;
 
+/// <summary>
+/// Basic type of any Git reference
+/// </summary>
 [Flags]
 public enum GitReferenceType
 {
+    /// <summary>
+    /// Invalid reference
+    /// </summary>
     Invalid = 0,
+    /// <summary>
+    /// A reference that points at an object id
+    /// </summary>
     Direct = 1,
+    /// <summary>
+    /// A reference that points at another reference
+    /// </summary>
     Symbolic = 2,
+    /// <summary>
+    /// Bitmask of all reference types
+    /// </summary>
     All = Direct | Symbolic
 }
 
-public unsafe readonly partial struct GitReference : IDisposable, IComparable<GitReference>
+/// <summary>
+/// In-memory representation of a reference
+/// </summary>
+/// <param name="nativeHandle">The native object pointer</param>
+[DebuggerDisplay("{NativeHandle == null ? (string?)null : GetName()}")]
+public unsafe readonly partial struct GitReference(Git2.Reference* nativeHandle) : IDisposable, IComparable<GitReference>, IGitHandle
 {
-    internal readonly Git2.Reference* NativeHandle;
+    /// <summary>
+    /// The native object pointer
+    /// </summary>
+    internal Git2.Reference* NativeHandle { get; } = nativeHandle;
 
-    internal GitReference(Git2.Reference* handle)
+    /// <inheritdoc/>
+    public bool IsNull => this.NativeHandle == null;
+
+    public void Dispose()
     {
-        NativeHandle = handle;
+        git_reference_free(this.NativeHandle);
     }
 
     public bool IsBranch
     {
         get
         {
-            var code = NativeApi.git_reference_is_branch(NativeHandle);
+            var handle = this.ThrowIfNull();
+
+            var code = git_reference_is_branch(handle.NativeHandle);
 
             return code != 0;
         }
@@ -41,7 +71,9 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     {
         get
         {
-            var code = NativeApi.git_reference_is_note(NativeHandle);
+            var handle = this.ThrowIfNull();
+
+            var code = git_reference_is_note(handle.NativeHandle);
 
             return code != 0;
         }
@@ -51,7 +83,9 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     {
         get
         {
-            var code = NativeApi.git_reference_is_remote(NativeHandle);
+            var handle = this.ThrowIfNull();
+
+            var code = git_reference_is_remote(handle.NativeHandle);
 
             return code != 0;
         }
@@ -61,35 +95,65 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     {
         get
         {
-            var code = NativeApi.git_reference_is_tag(NativeHandle);
+            var handle = this.ThrowIfNull();
+
+            var code = git_reference_is_tag(handle.NativeHandle);
 
             return code != 0;
         }
     }
 
-    internal byte* NativeName => NativeApi.git_reference_name(NativeHandle);
+    public bool IsSymbolic => this.NativeSymbolicTarget != null;
 
-    internal byte* NativeShorthand => NativeApi.git_reference_shorthand(NativeHandle);
-
-    internal byte* NativeSymbolicTarget => NativeApi.git_reference_symbolic_target(NativeHandle);
-
-    public GitRepository Owner => new(NativeApi.git_reference_owner(NativeHandle));
-
-    public GitReferenceType TargetType => NativeApi.git_reference_target_type(NativeHandle);
-
-    public void Dispose()
+    internal byte* NativeName
     {
-        NativeApi.git_reference_free(NativeHandle);
+        get
+        {
+            var handle = this.ThrowIfNull();
+
+            return git_reference_name(handle.NativeHandle);
+        }
+    }
+
+    internal byte* NativeSymbolicTarget
+    {
+        get
+        {
+            var handle = this.ThrowIfNull();
+
+            return git_reference_symbolic_target(handle.NativeHandle);
+        }
+    }
+
+    public GitRepository Owner
+    {
+        get
+        {
+            var handle = this.ThrowIfNull();
+
+            return new(git_reference_owner(handle.NativeHandle));
+        }
+    }
+
+    public GitReferenceType Type
+    {
+        get
+        {
+            var handle = this.ThrowIfNull();
+
+            return git_reference_type(handle.NativeHandle);
+        }
     }
 
     public override string ToString()
     {
-        return NativeHandle == null ? "< NULL >" : GetName();
+        return this.NativeHandle == null ? "< NULL HANDLE >" : GetName();
     }
 
+    /// <inheritdoc/>
     public int CompareTo(GitReference other)
     {
-        return NativeApi.git_reference_cmp(NativeHandle, other.NativeHandle);
+        return git_reference_cmp(this.NativeHandle, other.NativeHandle);
     }
 
     /// <summary>
@@ -100,25 +164,42 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     /// </remarks>
     public void Delete()
     {
-        Git2.ThrowIfError(NativeApi.git_reference_delete(NativeHandle));
+        var handle = this.ThrowIfNull();
+
+        Git2.ThrowIfError(git_reference_delete(handle.NativeHandle));
     }
 
     public GitReference Duplicate()
     {
-        GitReference reference;
-        Git2.ThrowIfError(NativeApi.git_reference_dup((Git2.Reference**)&reference, NativeHandle));
+        var handle = this.ThrowIfNull();
 
-        return reference;
+        Git2.Reference* reference = null;
+        Git2.ThrowIfError(git_reference_dup(&reference, handle.NativeHandle));
+
+        return new(reference);
     }
 
     public string GetName()
     {
-        return Utf8StringMarshaller.ConvertToManaged(this.NativeName)!;
+        var handle = this.ThrowIfNull();
+
+        return Git2.GetPooledString(git_reference_name(handle.NativeHandle));
+    }
+
+    public bool IsName(ReadOnlySpan<byte> utf8Name)
+    {
+        var handle = this.ThrowIfNull();
+
+        ReadOnlySpan<byte> name = MemoryMarshal.CreateReadOnlySpanFromNullTerminated(git_reference_name(handle.NativeHandle));
+
+        return name.SequenceEqual(utf8Name);
     }
 
     public string GetShorthand()
     {
-        return Utf8StringMarshaller.ConvertToManaged(this.NativeShorthand)!;
+        var handle = this.ThrowIfNull();
+
+        return Git2.GetPooledString(git_reference_shorthand(handle.NativeHandle));
     }
 
     /// <summary>
@@ -127,85 +208,145 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     /// <returns>The name of the symbolic target, or <see langword="null"/> if this isn't a symbolic reference</returns>
     public string? GetSymbolicTarget()
     {
-        return Utf8StringMarshaller.ConvertToManaged(this.NativeSymbolicTarget);
+        byte* target = this.NativeSymbolicTarget;
+
+        if (target is null)
+        {
+            return null;
+        }
+
+        return Git2.GetPooledString(target);
+    }
+
+    public GitObjectID? GetTarget()
+    {
+        var handle = this.ThrowIfNull();
+
+        var ptr = git_reference_target(handle.NativeHandle);
+
+        return ptr is null ? null : *ptr;
+    }
+
+    public GitObjectID? GetTargetPeel()
+    {
+        var handle = this.ThrowIfNull();
+
+        var ptr = git_reference_target_peel(handle.NativeHandle);
+
+        return ptr is null ? null : *ptr;
     }
 
     public GitObject Peel(GitObjectType type)
     {
-        Git2.Object* result;
-        Git2.ThrowIfError(NativeApi.git_reference_peel(&result, NativeHandle, type));
+        var handle = this.ThrowIfNull();
+
+        Git2.Object* result = null;
+        Git2.ThrowIfError(git_reference_peel(&result, handle.NativeHandle, type));
 
         return new(result);
     }
 
+    public TObject Peel<TObject>() where TObject: struct, IGitHandle, IGitObject<TObject>
+    {
+        var handle = this.ThrowIfNull();
+
+        Git2.Object* result = null;
+        Git2.ThrowIfError(git_reference_peel(&result, handle.NativeHandle, TObject.ObjectType));
+
+        return TObject.FromObjectPointer(result);
+    }
+
     public GitError Peel(GitObjectType type, out GitObject obj)
     {
-        Git2.Object* result;
-        var error = NativeApi.git_reference_peel(&result, NativeHandle, type);
+        var handle = this.ThrowIfNull();
+
+        Git2.Object* result = null;
+        var error = git_reference_peel(&result, handle.NativeHandle, type);
 
         obj = (error == GitError.OK) ? new(result) : default;
 
         return error;
     }
 
+    /// <summary>
+    /// Rename an existing reference.
+    /// </summary>
+    /// <param name="newName"></param>
+    /// <param name="force"></param>
+    /// <param name="logMessage"></param>
+    /// <returns></returns>
     public GitReference Rename(string newName, bool force, string? logMessage)
     {
-        GitReference newReference;
-        Git2.ThrowIfError(NativeApi.git_reference_rename((Git2.Reference**)&newReference, NativeHandle, newName, force ? 1 : 0, logMessage));
+        var handle = this.ThrowIfNull();
 
-        return newReference;
+        Git2.Reference* newReference = null;
+        Git2.ThrowIfError(git_reference_rename(&newReference, handle.NativeHandle, newName, force ? 1 : 0, logMessage));
+
+        return new(newReference);
     }
 
     public GitReference Resolve()
     {
-        GitReference peeledReference;
-        Git2.ThrowIfError(NativeApi.git_reference_resolve((Git2.Reference**)&peeledReference, NativeHandle));
+        var handle = this.ThrowIfNull();
 
-        return peeledReference;
+        Git2.Reference* peeledReference = null;
+        Git2.ThrowIfError(git_reference_resolve(&peeledReference, handle.NativeHandle));
+
+        return new(peeledReference);
     }
 
     public GitReference SetTarget(GitObjectID Id, string? logMessage)
     {
-        GitReference newReference;
-        Git2.ThrowIfError(NativeApi.git_reference_set_target((Git2.Reference**)&newReference, NativeHandle, &Id, logMessage));
+        var handle = this.ThrowIfNull();
 
-        return newReference;
+        Git2.Reference* newReference = null;
+        Git2.ThrowIfError(git_reference_set_target(&newReference, handle.NativeHandle, &Id, logMessage));
+
+        return new(newReference);
     }
 
     public GitReference SetTarget(in GitObjectID Id, string? logMessage)
     {
-        GitReference newReference;
+        var handle = this.ThrowIfNull();
+
+        Git2.Reference* newReference = null;
         GitError error;
 
         fixed (GitObjectID* ptr = &Id)
         {
-            error = NativeApi.git_reference_set_target((Git2.Reference**)&newReference, NativeHandle, ptr, logMessage);
+            error = git_reference_set_target(&newReference, handle.NativeHandle, ptr, logMessage);
         }
 
         Git2.ThrowIfError(error);
 
-        return newReference;
+        return new(newReference);
     }
 
     public GitReference SetSymbolicTarget(string target, string? logMessage)
     {
-        GitReference newReference;
-        Git2.ThrowIfError(NativeApi.git_reference_symbolic_set_target((Git2.Reference**)&newReference, NativeHandle, target, logMessage));
+        var handle = this.ThrowIfNull();
 
-        return newReference;
+        Git2.Reference* newReference = null;
+        Git2.ThrowIfError(git_reference_symbolic_set_target(&newReference, handle.NativeHandle, target, logMessage));
+
+        return new(newReference);
     }
 
     public GitReference SetSymbolicTarget(GitReference target, string? logMessage)
     {
-        GitReference newReference;
-        Git2.ThrowIfError(NativeApi.git_reference_symbolic_set_target((Git2.Reference**)&newReference, NativeHandle, target.NativeName, logMessage));
+        var handle = this.ThrowIfNull();
 
-        return newReference;
+        Git2.Reference* newReference = null;
+        Git2.ThrowIfError(git_reference_symbolic_set_target(&newReference, handle.NativeHandle, target.NativeName, logMessage));
+
+        return new(newReference);
     }
 
     public bool TryGetTarget(out GitObjectID id)
     {
-        var ptr = NativeApi.git_reference_target(NativeHandle);
+        var handle = this.ThrowIfNull();
+
+        var ptr = git_reference_target(handle.NativeHandle);
 
         if (ptr is not null)
         {
@@ -219,9 +360,11 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
 
     public bool TryGetTargetPeel(out GitObjectID id)
     {
-        var ptr = NativeApi.git_reference_target_peel(NativeHandle);
+        var handle = this.ThrowIfNull();
 
-        if (ptr is not null)
+        var ptr = git_reference_target_peel(handle.NativeHandle);
+
+        if (ptr != null)
         {
             id = *ptr;
             return true;
@@ -231,17 +374,35 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
         return false;
     }
 
+    public ref readonly GitObjectID DangerousGetTarget()
+    {
+        var handle = this.ThrowIfNull();
+
+        return ref *git_reference_target(handle.NativeHandle);
+    }
+
+    public ref readonly GitObjectID DangerousGetTargetPeel()
+    {
+        var handle = this.ThrowIfNull();
+
+        return ref *git_reference_target_peel(handle.NativeHandle);
+    }
+
     public static bool IsValidReferenceName(string referenceName)
     {
-        int code = NativeApi.git_reference_is_valid_name(referenceName);
+        return NormalizeReferenceName(null, referenceName, GitReferenceFormat.AllowOneLevel, throwIfInvalid: false);
+    }
 
-        return code != 0;
+    public static bool IsValidReferenceName(ReadOnlySpan<char> name)
+    {
+        return NormalizeReferenceName(null, name, GitReferenceFormat.AllowOneLevel, throwIfInvalid: false);
     }
 
     private static readonly SearchValues<char> _invalidChars = SearchValues.Create("~^:\\?[");
+    private static readonly SearchValues<string> _invalidSequences = SearchValues.Create(["..", "@{"], StringComparison.Ordinal);
 
     /// <summary>
-    /// Managed implementation of <see cref="NativeApi.git_reference_normalize_name(byte*, nuint, byte*, GitReferenceFormat)"/>.
+    /// Managed implementation of <see cref="git_reference_normalize_name(byte*, nuint, byte*, GitReferenceFormat)"/>.
     /// Only allows Ascii.
     /// </summary>
     /// <param name="referenceName"></param>
@@ -253,23 +414,69 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
     {
         ArgumentException.ThrowIfNullOrEmpty(referenceName);
 
+        var builder = new StringBuilder(Math.Max(referenceName.Length, 32));
+
+        bool success = NormalizeReferenceName(builder, referenceName, format, throwIfInvalid: true);
+        Debug.Assert(success);
+
+        return builder.ToString();
+    }
+
+    internal static bool NormalizeReferenceName(StringBuilder? builder, ReadOnlySpan<char> referenceName, GitReferenceFormat format, bool throwIfInvalid)
+    {
         const string GenericErrorMessage = "Provided reference name is invalid!";
 
         ReadOnlySpan<char> source = referenceName;
 
         if (!Ascii.IsValid(source)) // Potentially unnecessary, needs further research
-            ThrowInvalidSpec("Reference name must use only Ascii characters!");
+        {
+            if (throwIfInvalid)
+            {
+                ThrowInvalidSpec("Reference name must use only Ascii characters!"); 
+            }
+            else
+            {
+                return false;
+            }
+        }
 
         source = source.TrimStart('/');
 
         if (source[^1] == '/')
-            ThrowInvalidSpec("Reference name cannot end with '/'");
+        {
+            if (throwIfInvalid)
+            {
+                ThrowInvalidSpec("Reference name cannot end with '/'"); 
+            }
+            else
+            {
+                return false;
+            }
+        }
 
         if (source[^1] == '.')
-            ThrowInvalidSpec("Reference name cannot end with '.'");
+        {
+            if (throwIfInvalid)
+            {
+                ThrowInvalidSpec("Reference name cannot end with '.'"); 
+            }
+            else
+            {
+                return false;
+            }
+        }
 
         if (source.IndexOfAny(_invalidChars) >= 0 || source.IndexOfAnyInRange('\0', ' ') >= 0)
-            ThrowInvalidSpec("Reference name has invalid characters!");
+        {
+            if (throwIfInvalid)
+            {
+                ThrowInvalidSpec("Reference name has invalid characters!"); 
+            }
+            else
+            {
+                return false;
+            }
+        }
 
         int globCount = source.Count('*');
 
@@ -277,39 +484,75 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
         {
             if (globCount > 1)
             {
-                ThrowInvalidSpec("Reference name contains more than 1 glob! (A glob being '*')");
+                if (throwIfInvalid)
+                {
+                    ThrowInvalidSpec("Reference name contains more than 1 glob! (A glob being '*')"); 
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
         else
         {
             if (globCount > 0)
             {
-                ThrowInvalidSpec("Reference name is not allowed to be a glob pattern! (Not allowed to contain '*')");
+                if (throwIfInvalid)
+                {
+                    ThrowInvalidSpec("Reference name is not allowed to be a glob pattern! (Not allowed to contain '*')"); 
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
 
-        if (source.Contains("..", StringComparison.Ordinal)
-            || source.Contains("@{", StringComparison.Ordinal))
+        if (source.ContainsAny(_invalidSequences))
         {
-            ThrowInvalidSpec("Reference name cannot contain the sequences \"..\" or \"@{\"");
+            if (throwIfInvalid)
+            {
+                ThrowInvalidSpec("Reference name cannot contain the sequences \"..\" or \"@{\""); 
+            }
+            else
+            {
+                return false;
+            }
         }
 
         int segment_count = source.Count('/') + 1;
 
-        Span<Range> segments = segment_count > 16 ? new Range[segment_count] : stackalloc Range[segment_count];
+        Span<Range> segments = segment_count > 8 ? new Range[segment_count] : stackalloc Range[8];
 
         segment_count = source.Split(segments, '/', StringSplitOptions.RemoveEmptyEntries);
 
         if (segment_count == 1)
         {
             if ((format & GitReferenceFormat.AllowOneLevel) == 0)
-                ThrowInvalidSpec("'One level' reference names are not allowed!");
+            {
+                if (throwIfInvalid)
+                {
+                    ThrowInvalidSpec("'One level' reference names are not allowed!"); 
+                }
+                else
+                {
+                    return false;
+                }
+            }
 
             if ((format & GitReferenceFormat.RefSpecShorthand) == 0
                 && !(IsAllCapsAndUnderscore(source) || ((format & GitReferenceFormat.RefSpecPattern) != 0 && source.SequenceEqual("*"))))
             {
-                // TODO: Figure out what to say here
-                ThrowInvalidSpec(GenericErrorMessage);
+                if (throwIfInvalid)
+                {
+                    // TODO: Figure out what to say here
+                    ThrowInvalidSpec(GenericErrorMessage); 
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
         else // segment_count > 1
@@ -319,14 +562,19 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
             // The first segment is checked
             if (IsAllCapsAndUnderscore(source[segments[0]]))
             {
-                // TODO: Figure out what to say here
-                ThrowInvalidSpec(GenericErrorMessage);
+                if (throwIfInvalid)
+                {
+                    // TODO: Figure out what to say here
+                    ThrowInvalidSpec(GenericErrorMessage); 
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
 
         segments = segments[..segment_count];
-
-        var builder = new StringBuilder();
 
         for (int i = 0; i < segments.Length; ++i)
         {
@@ -334,21 +582,38 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
 
             if (segment.EndsWith(".lock"))
             {
-                ThrowInvalidSpec("Reference name segment cannot end with \".lock\"");
+                if (throwIfInvalid)
+                {
+                    ThrowInvalidSpec("Reference name segment cannot end with \".lock\"");
+                }
+                else
+                {
+                    return false;
+                }
             }
 
             if (segment[^1] == '.')
             {
-                ThrowInvalidSpec("Reference name segment cannot end with '.'");
+                if (throwIfInvalid)
+                {
+                    ThrowInvalidSpec("Reference name segment cannot end with '.'"); 
+                }
+                else
+                {
+                    return false;
+                }
             }
 
-            if (builder.Length > 0)
-                builder.Append('/');
+            if (builder is not null)
+            {
+                if (builder.Length > 0)
+                    builder.Append('/');
 
-            builder.Append(segment);
+                builder.Append(segment);
+            }
         }
 
-        return builder.ToString();
+        return true;
 
         static bool IsAllCapsAndUnderscore(ReadOnlySpan<char> span)
         {
@@ -411,5 +676,11 @@ public unsafe readonly partial struct GitReference : IDisposable, IComparable<Gi
         {
             throw new ArgumentException(message, nameof(referenceName));
         }
+    }
+
+    internal static void ThrowIfInvalidReferenceName(ReadOnlySpan<char> referenceName, GitReferenceFormat format)
+    {
+        bool success = NormalizeReferenceName(null, referenceName, format, throwIfInvalid: true);
+        Debug.Assert(success);
     }
 }
