@@ -13,16 +13,19 @@ public unsafe struct GitMessageTrailerArray
     public GitMessageTrailer* Trailers;
     public nuint Count;
 
+#pragma warning disable IDE0044, IDE0051, CS0169 // Add readonly modifier
+    // Relevent to the native library, DO NOT REMOVE.
     private byte* _trailer_block;
+#pragma warning restore IDE0044, IDE0051, CS0169 // Add readonly modifier
 
-    public KeyValuePair<string, string>[] ToManaged()
+    public readonly KeyValuePair<string, string>[] ToManaged()
     {
         ReadOnlySpan<GitMessageTrailer> span = new(Trailers, checked((int)Count));
 
         var array = new KeyValuePair<string, string>[span.Length];
         for (int i = 0; i < span.Length; ++i)
         {
-            var key = Utf8StringMarshaller.ConvertToManaged(span[i].Key);
+            var key = Git2.GetPooledString(span[i].Key);
             var value = Utf8StringMarshaller.ConvertToManaged(span[i].Value);
 
             array[i] = new(key!, value!);
