@@ -1,12 +1,12 @@
 ﻿namespace SharpGit2
 {
-    public unsafe struct GitRemoteHead
+    public unsafe struct GitRemoteHead(in Native.GitRemoteHead native)
     {
-        public bool Local;
-        public GitObjectID OID;
-        public GitObjectID LOID;
-        public string Name;
-        public string SymRefTarget;
+        public bool Local = native.Local != 0;
+        public GitObjectID OID = native.OID;
+        public GitObjectID LOID = native.LOID;
+        public string Name = Git2.GetPooledString(native.Name)!;
+        public string SymRefTarget = Git2.GetPooledString(native.SymRefTarget)!;
     }
 }
 
