@@ -1,4 +1,6 @@
-﻿using System.Runtime.InteropServices.Marshalling;
+﻿using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.Marshalling;
+using System.Text;
 
 namespace SharpGit2.Native;
 
@@ -7,16 +9,34 @@ public unsafe struct GitStringArray(byte** strings, nuint count)
     public byte** Strings = strings;
     public nuint Count = count;
 
-    public readonly string[] ToManaged()
+    public readonly string[] ToManaged(bool poolValues = false)
     {
         int count = checked((int)Count);
         byte** ptr = Strings;
 
+        if (count == 0 || ptr == null)
+            return [];
+
         var managedArray = new string[count];
 
-        for (int i = 0; i < count; ++i)
+        if (poolValues)
         {
-            managedArray[i] = Utf8StringMarshaller.ConvertToManaged(ptr[i])!;
+            for (int i = 0; i < count; ++i)
+            {
+                var nativeValue = ptr[i];
+
+                if (nativeValue != null)
+                {
+                    managedArray[i] = Git2.GetPooledString(nativeValue);
+                }
+            }
+        }
+        else
+        {
+            for (int i = 0; i < count; ++i)
+            {
+                managedArray[i] = Utf8StringMarshaller.ConvertToManaged(ptr[i])!;
+            }
         }
 
         return managedArray;
